fix(epistemic): separate proof-unit scanner concerns
